implemented support for remote DCS install config
DCS Locations to Update can now be set to know that this is a remote Helios
machine and not the one running DCS.   This disables configuration checks and
dialogs that do not apply in such a scenario.  This option is global in Helios settings
and applies to all profiles on the local machine.