dev tools:  Simulate monitor layout stored in current profile
This tool can be used to reset profiles to the monitors store in some profile.   This can be used to
reset profiles to a reference configration like Capt Zeen's monitor layout or to reproduce reset
monitor problems observed by a user by setting the simulated monitors from a new profile or
reset profile provided by the user.

Simulated monitors are discarded on close of Profile Editor or when the "Stop simulating" tools
menu item is selected.