#605 Hornet UFC 16 Segment display Issues (#608)
* Bump the DCSInterfaces submodule

* Fix for #605 which was a fix for the Hornet UFC 16 digit displays.  The "12" did not appear due to it being mapped to a soft hyphen charater. and the scratchpad 16 digit displays did not correctly show numeric values greater than 20.  This seems to just be an issue when precision entry is selected.

The font for the Hornet UFC has been updated to include left and right aligned numbers.  The UFC has also been updated to allow the 16 segment displays to be formed from two different characters.

Note: DCS sends different characters in the tens position depending on what the units position value is   This seems to be related to the combined central segments of the 16 segment display, however it does not appear to be exactly  as I would anticipate.  For this reason, teh dictionary contains multiple conversions for the same tens digit.

This commit covers 1.6 of the Hornet UFC font, and changes to enable dual digits up to 99 in the 16 segment displays.