Submodule update for AV-8B interface
AV-8B Interface change to Correct V/UHF Operational Mode switch

Changes to AV-8B Interface emergency codes 447 and 470

Addditional test added to switch to do a numerical comparions of the argument value and the netwrok value.  This was needed because a switch with %0.2f was failing a string test for 0 because the argument value for the zero position was 0.0 and not 0.00

AV-8B Interface changes to resolve several multi-position switches as well as Radio Channel selector binding change.