Timer Panel - Timer Actions for Enable and Interval (#613)
* Creates Action to Enable / Disable the timer function and also two actions to set the timer interval- on temporarily and the other to sett the default.  When the panel is hidden, the interval is always set to the value configured for the default interval whether that is in in PE or by the profile setting a new default.  This is to ensure the repeatability of activity on the timer panel.
Use Case:  A timer panel is made visible with a control for the landing gear handle on it expecting that it might be interacted with, and a long time interval would be configured for the panel so that it would eventually hide if the user ignores it and does not lower the Landing gear.  If the user drops the landing gear, then there is the possibility that the gear might be retracted again so the panel is not hidden immediately, but a shorter interval is set for the panel to give the pilot a chance to confirm the gear is down, but if no more input for the new shorter period, the panel will hide itself.  There could also be a button on the panel for the user to disable the timer so that it does not auto hide.

* Fixed problem where the timer was disabled in profile editor, then enabled by a control in the profile, but no timer was created.

* Fix for issue where the timer is disabled in PE, the panel is hidden, and the interval is set via action,.  In this situation _timer would not have been created.

* Debug logging added to Timer Panel, more adjustments to when the Timer is created, and stopiing the Time Out value being dependent on whether the timer is enabled fir the control behaviour panel.

* Creates Action to Enable / Disable the timer function and also an action to set the timer interval.  When the panel is un-hidden, the interval is always set to the value configured for the Timer Panel in PE to ensure the repeatability of activity on the timer panel.
Use Case:  A timer panel is made visible with a control for the landing gear handle on it expecting that it might be interacted with, and a long time interval would be configured for the panel so that it would eventually hide if the user ignores it and does not lower the Landing gear.  If the user drops the landing gear, then there is the possibility that the gear might be retracted again so the panel is not hidden immediately, but a shorter interval is set for the panel to give the pilot a chance to confirm the gear is down, but if no more input for the new shorter period, the panel will hide itself.  There could also be a button on the panel for the user to disable the timer so that it does not auto hide.

Debug logging added to Timer Panel, more adjustments to when the Timer is created, and stopiing the Time Out value being dependent on whether the timer is enabled fir the control behaviour panel.

Fix for issue where the timer is disabled in PE, the panel is hidden, and the interval is set via action,.  In this situation _timer would not have been created.

Fixed problem where the timer was disabled in profile editor, then enabled by a control in the profile, but no timer was created.

* Redesign to (hopefully) reduce confusion.
Now has one action to set the default interval, and a second for a temporary interval to be used until the panel hides.
When the panel hides, the interval is set to the default interval value which is initially retrieved from the profile but can be update via the action to set the default interval.

* Correction of words for Once Only Interval