Fixing copyright date on SimpleComputeWire.cs
Adding some bounds checking when setting the thresholds on the Simple Compute Wire.

Removal of the "-- changes to this viewport are not tracked automatically." comments which actually caused the lua file to be detected as changed.

Simple Calculation Wire control.  Move of many of the "special" controls from the Toolbox "Miscellaneous" section to "Special Controls".